Added a bit of a code skeleton.
We've got DBReader for the PQXX connection to a database. More specific functionality could be added when the PQXX website comes back online and we know exactly what format SCIP would like its data. There's Optimizer for storing SCIP variables and constraints, then actually running the solver. Finally, Scheduler acts as a mediator between them and as a facade for easy control of the whole system.