Made Scheduler obey Object Calisthenics better.
The only real violations I see are having three member variables (but it would be terrible design to bring it down to two, so I'm happy with it) and having ScheduleData functions that return the entire vectors. I didn't really feel like completely changing Optimizer to support it, and it's not too big a deal anyway. At least modification is still restricted.